+++ Application/Utils.cpp
@@ -41,7 +41,7 @@ namespace Utils
 
 	}
 
-	void CalcHist(uint8_t* scan0, uint32_t stride, int w, int h, std::vector<int> &histR, std::vector<int> &histG, std::vector<int> &histB, std::vector<int> &histA) {
+	void CalcHist(uint32_t* scan0, uint32_t stride, int w, int h, std::vector<int> &histR, std::vector<int> &histG, std::vector<int> &histB, std::vector<int> &histA) {
 
 		uint32_t *pLine = (uint32_t*)((uint8_t *)scan0);
